Protect critical data.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.06@2040 0a995953-517f-4c62-b348-104708d926b2

+++ GUI/spot-on-a.cc
@@ -2684,7 +2684,8 @@ void spoton::slotDeactivateKernel(void)
 		      0,
 		      &libspotonHandle,
 		      m_settings.value("gui/gcryctl_init_secmem",
-				       65536).toInt()) == LIBSPOTON_ERROR_NONE)
+				       65536).
+		      toInt()) == LIBSPOTON_ERROR_NONE)
     libspoton_deregister_kernel
       (libspoton_registered_kernel_pid(&libspotonHandle, 0),
        &libspotonHandle);
